Add Wayland display support
This allows talking to Wayland compositors directly, rather than relying
on XWayland data which often doesn't reflect the exact setup on Wayland,
and of course doesn't work without XWayland running.

It tries to use xdg_output for precise data, but falls back to wl_output
information if that protocol is missing.