Update HS64 configuration
The DeviceString for index 201 for HS64 appends some sort of revision number, so it diverges from the usual format of `HUION_Txxx` where xxx is the company's internal codenumber for the model.

HS64's DeviceString index 201 have a format of `HUION_Txxx_yyyyyy` where yyyyyy can vary from device to device of the same HS64 model.