Add support for additional tablet report data
Tilt is supported for intuosv2, v3 and in generic tablets with
long enough report size to fit tilt.
Proximity is supported for intuosv2 and v3, not sure where
it is on other wacoms
Eraser detection rn is only for intuosv3, v2 supposedly identifies
eraser by tool id packet, that's not yet supported.