BindingHandler: Rewrite pressure reports to account for thresholds.
`IInterruptBinding`s are currently invoked regardless of state, and has no
way of knowing the pressure threshold, resulting in pressure threshold
being useless for some bindings, such as PenPassthroughBinding.

By rewriting the pressure in the report we achieve correct behavior in
`EvdevVirtualTablet`'s `SetPressure()` and allows tip and eraser
thresholds to be effective even when using `IInterruptBinding`s.

This also has the additional side effect of only triggering `BTN_TOUCH`
events in `EvdevVirtualTablet` when the pressure exceeds the users
specified threshold, which would previously trigger incorrectly on
tablets with noisy pressure reports.