Allow more report ids
This will allow more report ids. The previous filter was a little over-restrictive, as discovered in discord.
This will attempt to parse information, if any bit is not zero, excluding the leading 20 and excluding the last 2 bytes (last byte is scroll-weel, and the one before that is button information).
This should be enough restrictions, from what I can gather.