Add Entry for Artist 16 2nd Gen
I have only tested the config file on the public stable build that only supports 9 tablet hotkey buttons, I am just assuming the 10th will start working with the update. It shows up in the tablet debugger so I have no reason to expect that it would not but it is worth noting.