Fix UC-Logic 1060N configuration I don't think we had concrete evidence of this configuration working to begin with.
No verification was ever done on pressure and considering the age of
this tablet I'm guessing it was plain wrong to begin with, so I've
corrected the value to 2047. See #1037 for the original implementation.

Somewhere between the original pull request mentioned above and now the
`OutputReportLength` was set to null, which is wrong because multiple
interfaces have the `InputReportLength` of 8, which at a minimum breaks
support for Linux and possibly MacOS.