UX: Rewrite FileDialog directory checks
Previously, we assumed that Eto.EtoEnvironment.GetFolderPath would never
return an empty string. It is however apparent that this is actually the
case on some systems, seen on both Windows and Linux.

Constructing URI's with an empty string throws an exception, which this
commit attempts to avoid. It was not an option to simply change the Uri
constructor, as Directory can only be set if it is a valid Uri, so
instead the Save/OpenFileDialog Eto constructor has been moved to an
extension method internal to the project.