[0.6.x] macOS: Track modifier status in OTD instead of fetching from system
CGEventSourceFlagsState does not update flags immediately after an event is posted,
leading to unreliable shortcut behavior. This commit modifies the OTD to track
modifier status within OTD itself. Since it no longer fetches modifier flags from
the system, it cannot combine with other keyboards to form shortcuts. However, this
is likely rare, as it is already not possible on macOS to combine keys on different
keyboards to form shortcuts.

fixes #3889